pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Hadi Aji <hadeaji@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
django-heroku = "^0.3.1"
django-environ = "^0.4.5"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
djangorestframework-simplejwt = "^4.6.0"
psycopg2 = "^2.8.6"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
